% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b3_indexreport.R
\name{b3_indexreport_download}
\alias{b3_indexreport_download}
\title{Get Index Report BVBG.087.01}
\usage{
b3_indexreport_download(date, last = TRUE, directory = tempdir())
}
\arguments{
\item{date}{"yyyy-mm-dd" in class Date format}

\item{last}{logical setting if we keep all files or choose the last one}

\item{directory}{a string for the directory where the files will be saved}
}
\value{
a file
}
\description{
Downloads archive Index Report BVBG.087.01 from B3's website
}
\examples{
\donttest{
b3_indexreport_download(as.Date("2020-01-31"))
}
}
