% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b3_indexreport.R
\name{b3_indexreport_read}
\alias{b3_indexreport_read}
\title{Read and process Index Report BVBG.087.01}
\usage{
b3_indexreport_read(file)
}
\arguments{
\item{file}{path to the file}
}
\value{
a data_frame with tidy data
}
\description{
Read and process Index Report BVBG.087.01
}
\examples{
\donttest{
b3_indexreport_read(b3_indexreport_download(as.Date("2020-01-31")))
}
}
